refactor(flynn): GNB 내 존재했던 여러 컴포넌트를 분리
