+++ src/components/GNB/GNB.module.css
@@ -27,40 +27,6 @@
   height: 2.4rem;
 }
 
-/* 로그인 전 */
-.loginBtn {
-  width: 12.8rem;
-  height: 5.3rem;
-  border-radius: 0.8rem;
-  font-size: 1.8rem;
-  font-weight: 600;
-  line-height: 5.3rem;
-  text-align: center;
-  background: linear-gradient(90.99deg, var(--primary) 0.12%, #6ae3fe 101.84%);
-  color: #f5f5f5;
-}
-.loginBtn:hover {
-  cursor: pointer;
-}
-
-/* 로그인 후 */
-.userContainer {
-  display: flex;
-  align-items: center;
-  gap: 0.6rem;
-  color: var(--gray-brt1);
-}
-
-.userImg {
-  width: 2.8rem;
-  height: 2.8rem;
-  border-radius: 100%;
-}
-
-.userEmail {
-  font-size: 1.4rem;
-}
-
 /* Tablet 반응형 */
 @media (max-width: 1200px) {
   .gnb {
@@ -89,18 +55,5 @@
   .logo {
     width: 8.867rem;
     height: 1.6rem;
-  }
-
-  .loginBtn {
-    width: 8rem;
-    height: 3.7rem;
-    border-radius: 0.8rem;
-    font-size: 1.4rem;
-    font-weight: 600;
-    line-height: 3.7rem;
-  }
-
-  .userEmail {
-    display: none;
   }
 }
