+++ src/components/GNB/LoginButton.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+import { Link } from "react-router-dom";
+import styles from "./LoginButton.module.css";
+
+function LoginButton() {
+  return (
+    <Link to="/signin" className={styles.loginBtn}>
+      로그인
+    </Link>
+  );
+}
+
+export default LoginButton;
