+++ src/components/GNB/LoginButton.module.css
@@ -1,0 +1,27 @@
+.loginBtn {
+  width: 12.8rem;
+  height: 5.3rem;
+  border-radius: 0.8rem;
+  font-size: 1.8rem;
+  font-weight: 600;
+  line-height: 5.3rem;
+  text-align: center;
+  background: linear-gradient(90.99deg, var(--primary) 0.12%, #6ae3fe 101.84%);
+  color: #f5f5f5;
+}
+
+.loginBtn:hover {
+  cursor: pointer;
+}
+
+/* Mobile 반응형 */
+@media (max-width: 768px) {
+  .loginBtn {
+    width: 8rem;
+    height: 3.7rem;
+    border-radius: 0.8rem;
+    font-size: 1.4rem;
+    font-weight: 600;
+    line-height: 3.7rem;
+  }
+}
